add very basic transaction system
This changes allows to commit atomically the changes to one or more
btree at the same time.

A transaction is handled via a function sent to the updater
asynchronously, the process requesting the transaction is waiting for
its result.

Btree functions running in the function use the current database value
handled in the process dictionnary. Only one transaction can run at a
time but multiple concurrent transaction can be requested. They are then
queued before a commit happen.

Until the header is commited the database value isn't changed.

Ex of a transaction incrementing atomically a value in the database:

    {ok, NewCount} = cbt:transact(Db, fun() ->
                    [{ok, {Key, Counter}}] = cbt:lookup(test, [Key]),
                    NewCounter1 = Counter + 1,
                    ok = cbt:add_remove(test, [{Key, NewCounter1}], []),
                    {ok, NewCounter1}
            end),