add cbt_btree_copy
This module copy a btree to another file. It is based on the couchbase
implementation from Filipe. It works as follow:

1) Extract each key-value pair from the original btree and add it
to the leaf key-value list maintained in memory.
2) When the leaf list's size reaches to the chunk threshold, it
is appended into the new btree file as a leaf node. And then,
add its offset to the parent node maintained in memory. If the
parent node's size also reaches to the threshold, it is appended
into the btree file as well. This process is repeated until there
is no non-leaf node to be appended into the btree file.
3) Repeat the above steps until the last key-value pair is read.

Although this approach creates a nearly complete btree, it
does not require an additional full scanning from the new file to
construct a complete btree. Regarding the memory overhead, it
only requires memory that can hold one node at each btree level
at any given time.